+++ CHANGELOG.rst
@@ -21,6 +21,7 @@ Added
     - ``syncstudytables`` management command (#1509)
     - ``get_last_material_index()`` helper (#1554)
     - ``get_latest_file_path()`` helper (#1554)
+    - "Not found" element for iRODS modal filter (#1562)
 
 Changed
 -------